requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
django-allauth==0.42.0
django-filter==2.3.0
Pillow==7.2.0
gunicorn==20.0.4
whitenoise==5.1.0